Rename package circuit_breaker to circuitbreaker.
